extract useChatService()
